Python code and Dataset
This dataset is also available on https://www.kaggle.com/c/prudential-life-insurance-assessment/data. You can also check the kernel for how people approach to this problem for increasing the accuracy of the prediction. I used this dataset to validate the backward elimination algorithm and Adjusted R squared trick to narrow down the focus on important features of the life insurance industry.